chore: extract clap_enum_for_named_arg module
