+++ interactive-clap-derive/src/derives/interactive_clap/mod.rs
@@ -25,52 +25,8 @@ pub fn impl_interactive_clap(ast: &syn::DeriveInput) -> TokenStream {
             let to_interactive_clap_context_scope_trait_block =
                 self::structs::to_interactive_clap_context_scope_trait::token_stream(ast, &fields);
 
-            let clap_enum_for_named_arg = fields.iter().find_map(|field| {
-                let ident_field = &field.clone().ident.expect("this field does not exist");
-                let variant_name_string = crate::helpers::snake_case_to_camel_case::snake_case_to_camel_case(ident_field.to_string());
-                let variant_name = &syn::Ident::new(&variant_name_string, Span::call_site());
-                let attr_doc_vec: Vec<_> = field.attrs.iter()
-                    .filter(|attr| attr.path.is_ident("doc"))
-                    .map(|attr| attr.into_token_stream())
-                    .collect();
-                field.attrs.iter()
-                    .filter(|attr| attr.path.is_ident("interactive_clap"))
-                    .flat_map(|attr| attr.tokens.clone())
-                    .filter(|attr_token| {
-                        match attr_token {
-                            proc_macro2::TokenTree::Group(group) => group.stream().to_string() == *"named_arg",
-                            _ => abort_call_site!("Only option `TokenTree::Group` is needed")
-                        }
-                    })
-                    .map(|_| {
-                        let ty = &field.ty;
-                        let type_string = match ty {
-                            syn::Type::Path(type_path) => {
-                                match type_path.path.segments.last() {
-                                    Some(path_segment) => path_segment.ident.to_string(),
-                                    _ => String::new()
-                                }
-                            },
-                            _ => String::new()
-                        };
-                        let enum_for_clap_named_arg = syn::Ident::new(&format!("ClapNamedArg{}For{}", &type_string, &name), Span::call_site());
-                        quote! {
-                            #[derive(Debug, Clone, clap::Parser, interactive_clap_derive::ToCliArgs)]
-                            pub enum #enum_for_clap_named_arg {
-                                #(#attr_doc_vec)*
-                                #variant_name(<#ty as interactive_clap::ToCli>::CliVariant)
-                            }
-
-                            impl From<#ty> for #enum_for_clap_named_arg {
-                                fn from(item: #ty) -> Self {
-                                    Self::#variant_name(<#ty as interactive_clap::ToCli>::CliVariant::from(item))
-                                }
-                            }
-                        }
-                    })
-                    .next()
-                })
-                .unwrap_or(quote!());
+            let clap_enum_for_named_arg_block =
+                self::structs::clap_enum_for_named_arg::token_stream(ast, &fields);
 
             quote! {
                 #to_cli_trait_block
@@ -81,7 +37,7 @@ pub fn impl_interactive_clap(ast: &syn::DeriveInput) -> TokenStream {
 
                 #from_cli_trait_block
 
-                #clap_enum_for_named_arg
+                #clap_enum_for_named_arg_block
             }
         }
         syn::Data::Enum(syn::DataEnum { variants, .. }) => {
@@ -238,6 +194,9 @@ mod structs {
 
     #[doc = include_str!("../../../docs/structs_from_cli_trait_docstring.md")]
     pub mod from_cli_trait;
+
+    #[doc = include_str!("../../../docs/clap_enum_for_named_arg_docstring.md")]
+    pub mod clap_enum_for_named_arg;
 
     /// these are common field methods, reused by other [structs](super::structs) submodules
     pub(super) mod common_field_methods;
